requirements.txt
python-docx==1.1.0
pdfminer.six==20221105
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.4.0
pytesseract==0.3.10